pandas basics
I have a daily csv emailed to me that I save to a folder. 
I use glob & os.path.getctime to pulll the latest csv from that folder.
This file holds some of the base actions I can use on that file, depending on what I want to do.
The basics of the csv is a list of devices that report to a server the number of messages it still has to process, and within each message there are x number of frames.